third_party/blink/web_tests/external/wpt/html/syntax/parsing/html5lib_write_single_file=html5test-com-expected.txt

This is a testharness.js-based test.
[FAIL] <?import namespace="foo" implementation="#bar">
  assert_equals: expected "#document\\n| <!-- ?import namespace=\\"foo\\" implementation=\\"#bar\\" -->\\n| <html>\\n|   <head>\\n|   <body>" but got "#document\\n| <html>\\n|   <head>\\n|   <body>"
Harness: the test ran to completion.
